+++ src/utils/utils.js
@@ -1,7 +1,7 @@
 import chalk from "chalk";
 import figlet from "figlet";
-import shell from "shelljs";
 import pressAnyKey from "press-any-key";
+import shell from "shelljs";
 import menu from "../cli.js";
 
 export const helpMe = () => {
@@ -11,19 +11,12 @@ export const helpMe = () => {
   const pipenv = chalk.cyan.bold("Pipenv");
   const vsc = chalk.cyan.bold("Visual Studio Code");
 
-  shell.echo();
-  shell.echo(`${frames} scaffolds apps in different flavors.`);
-  shell.echo();
-  shell.echo(`If ${vsc} is installed, it starts it once the all is done.`);
-  shell.echo();
-  shell.echo(`${frames} checks for ${pipenv} and eventually installs it.`);
-  shell.echo();
-  shell.echo(`Run ${frames} in the parent directory of your app.`);
-  shell.echo();
-  shell.echo(`${frames} was created by ${author}`);
-  shell.echo();
-  shell.echo(`Feel free to use, share and contribute: ${url}`);
-  shell.echo();
+  shell.echo(`\n${frames} scaffolds apps in different flavors.`);
+  shell.echo(`\nIf ${vsc} is installed, it starts it once the all is done.`);
+  shell.echo(`\n${frames} checks for ${pipenv} and eventually installs it.`);
+  shell.echo(`\nRun ${frames} in the parent directory of your app.`);
+  shell.echo(`\n${frames} was created by ${author}`);
+  shell.echo(`\nFeel free to use, share and contribute: ${url}\n`);
 
   pressAnyKey(
     chalk.greenBright("Press any key to go back to the menu...")
@@ -32,25 +25,26 @@ export const helpMe = () => {
   });
 };
 
-export const goodbye = (name) => {
-  const path = chalk.bgBlue.whiteBright(` ${process.cwd()}/${name} `);
+export const goodbye = (framework, framecolor, name) => {
   if (!name) {
     shell.echo();
-    shell.echo(chalk.yellowBright.bold(figlet.textSync("Bye!")));
+    shell.echo(chalk.yellowBright.bold(figlet.textSync("  Bye!")));
     shell.echo();
     return;
   }
-  console.log();
-  console.log(chalk.greenBright.bold("All done!"));
-  console.log();
-  console.log(
+  const path = framecolor(` ${process.cwd()}/${name} `);
+  const project = `${framecolor(` ${framework} `)} project`;
+  shell.echo();
+  shell.echo(chalk.greenBright.bold("All done!"));
+  shell.echo();
+  shell.echo(
     chalk.cyanBright.bold(
-      `Your frame is ready at ${path}. Now go and create a great app!`
+      `Your ${project} is ready at ${path}. Now go and create a great app!`
     )
   );
-  console.log();
-  console.log(chalk.yellowBright.bold(figlet.textSync("Bye!")));
-  console.log();
+  shell.echo();
+  shell.echo(chalk.yellowBright.bold(figlet.textSync("  Bye!")));
+  shell.echo();
 };
 
 export const noPython = () => {